:recycle: [Refactor]: fSD 구조 적용 - 1. 코드를 pages 단위로 분할
