+++ apps/client/src/pages/Home/HomePage.tsx
@@ -1,0 +1,11 @@
+import LiveList from '@pages/Home/LiveList';
+import Banner from './Banner';
+
+export function HomePage() {
+  return (
+    <div className="flex flex-col justify-start w-full min-h-[calc(100vh-74px)]">
+      <Banner />
+      <LiveList />
+    </div>
+  );
+}
